=============== ALL FRESHMEN PLEASE NOTE ===============
To minimize scheduling confusion, please realize that if you are taking one
course which is offered at only one time on a given day, and another which is
offered at all times on that day, the second class will be arranged as to
afford maximum inconvenience to the student.  For example, if you happen
to work on campus, you will have 1-2 hours between classes.  If you commute,
there will be a minimum of 6 hours between the two classes.