Watch Rincewind.
Look at him.  Scrawny, like most wizards, and clad in a dark red robe on
which a few mystic sigils were embroidered in tarnished sequins. Some might
have taken him for a mere apprentice enchanter who had run away from his
master out of defiance, boredom, fear and a lingering taste for
heterosexuality.  Yet around his neck was a chain bearing the bronze octagon
that marked him as an alumnus of Unseen University, the high school of magic
whose time-and-space transcendent campus is never precisely Here or There.
Graduates were usually destined for mageship at least, but Rincewind--after
an unfortunate event--had left knowing only one spell and made a living of
sorts around the town by capitalizing on an innate gift for languages.  He
avoided work as a rule, but had a quickness of wit that put his
acquaintances in mind of a bright rodent.
		-- Terry Pratchett, \The Colour of Magic\