If men are not afraid to die, it is of no avail to threaten them with death.
If men live in constant fear of dying,
And if breaking the law means a man will be killed,
Who will dare to break the law?

There is always an official executioner.
If you try to take his place,
It is like trying to be a master carpenter and cutting wood.
If you try to cut wood like a master carpenter,
	you will only hurt your hand.
		-- Tao Te Ching, \Lao Tsu, #74\