First, a few words about tools.
Basically, a tool is an object that enables you to take advantage of the
laws of physics and mechanics in such a way that you can seriously injure
yourself.  Today, people tend to take tools for granted.  If you\'re ever
walking down the street and you notice some people who look particularly
smug, the odds are that they are taking tools for granted.  If I were you,
I\'d walk right up and smack them in the face.
		-- Dave Barry, \The Taming of the Screw\