THE STORY OF CREATION 			       or 			 THE MYTH OF URK
In the beginning there was data.  The data was without form and null, and
darkness was upon the face of the console; and the Spirit of IBM was moving
over the face of the market.  And DEC said, \Let there be registers;\ and
there were registers.  And DEC saw that they carried; and DEC separated the
data from the instructions.  DEC called the data Stack, and the instructions
they called Code.  And there was evening and there was morning, one interrupt
...
		-- Rico Tudor