This is Jim Rockford. At the tone leave your name and message; I\'ll get back to you.
This is Maria, Liberty Bail Bonds.  Your client, Todd Lieman, skipped and
his bail is forfeit.  That\'s the pink slip on your \'74 Firebird, I believe.
Sorry, Jim, bring it on over.

This is Marilyn Reed, I wanta talk to you...  Is this a machine?  I don\'t
talk to machines\!  [Click]
		-- \The Rockford Files\