Between 1950 and 1952, a bored weatherman, stationed north of Hudson Bay, left a monument that neither government nor time can eradicate. Using a bulldozer abandoned by the Air Force, he spent two years and great effort pushing boulders into a single word.
It can be seen from 10,000 feet, silhouetted against the snow.
Government officials exchanged memos full of circumlocutions (no Latin
equivalent exists) but failed to word an appropriation bill for the
destruction of this cairn, that wouldn\'t alert the press and embarrass
both Parliament and Party.

It stands today, a monument to human spirit.  If life exists on other
planets, this may be the first message received from us.
		-- The Realist, November, 1964.