If I have seen farther than others, it is because I was standing on the shoulders of giants. 		-- Isaac Newton
In the sciences, we are now uniquely priviledged to sit side by side with
the giants on whose shoulders we stand.
		-- Gerald Holton

If I have not seen as far as others, it is because giants were standing on
my shoulders.
		-- Hal Abelson

Mathematicians stand on each other\'s shoulders.
		-- Gauss

Mathemeticians stand on each other\'s shoulders while computer scientists
stand on each other\'s toes.
		-- Richard Hamming

It has been said that physicists stand on one another\'s shoulders.  If
this is the case, then programmers stand on one another\'s toes, and
software engineers dig each other\'s graves.
		-- Unknown