... an anecdote from IBM\'s Yorktown Heights Research Center.  When a programmer used his new computer terminal, all was fine when he was sitting down, but he couldn\'t log in to the system when he was standing up.  That behavior was 100 percent repeatable: he could always log in when sitting and never when standing.
Most of us just sit back and marvel at such a story; how could that terminal
know whether the poor guy was sitting or standing?  Good debuggers, though,
know that there has to be a reason.  Electrical theories are the easiest to
hypothesize: was there a loose with under the carpet, or problems with static
electricity?  But electrical problems are rarely consistently reproducible.
An alert IBMer finally noticed that the problem was in the terminal\'s keyboard:
the tops of two keys were switched.  When the programmer was seated he was a
touch typist and the problem went unnoticed, but when he stood he was led
astray by hunting and pecking.
	-- \Programming Pearls\ column, by Jon Bentley in CACM February 1985