I made it a rule to forbear all direct contradictions to the sentiments of others, and all positive assertion of my own.  I even forbade myself the use of every word or expression in the language that imported a fixed opinion, such as \certainly\, \undoubtedly\, etc.   I adopted instead of them \I conceive\, \I apprehend\, or \I imagine\ a thing to be so or so; or \so it appears to me at present\.
When another asserted something that I thought an error, I denied myself the
pleasure of contradicting him abruptly, and of showing him immediately some
absurdity in his proposition.  In answering I began by observing that in
certain cases or circumstances his opinion would be right, but in the present
case there appeared or semed to me some difference, etc.

I soon found the advantage of this change in my manner; the conversations I
engaged in went on more pleasantly.  The modest way in which I proposed my
opinions procured them a readier reception and less contradiction.  I had
less mortification when I was found to be in the wrong, and I more easily
prevailed with others to give up their mistakes and join with me when I
happened to be in the right.
-- Autobiography of Benjamin Franklin