Oxygen is a very toxic gas and an extreme fire hazard.  It is fatal in concentrations of as little as 0.000001 p.p.m.  Humans exposed to the oxygen concentrations die within a few minutes.  Symptoms resemble very much those of cyanide poisoning (blue face, etc.).  In higher concentrations, e.g. 20%, the toxic effect is somewhat delayed and it takes about 2.5 billion inhalations before death takes place.  The reason for the delay is the difference in the mechanism of the toxic effect of oxygen in 20% concentration.  It apparently contributes to a complex process called aging, of which very little is known, except that it is always fatal.
However, the main disadvantage of the 20% oxygen concentration is in the
fact it is habit forming.  The first inhalation (occurring at birth) is
sufficient to make oxygen addiction permanent.  After that, any
considerable decrease in the daily oxygen doses results in death with
symptoms resembling those of cyanide poisoning.

Oxygen is an extreme fire hazard.  All of the fires that were reported in
the continental U.S. for the period of the past 25 years were found to be
due to the presence of this gas in the atmosphere surrounding the buildings
in question.

Oxygen is especially dangerous because it is odorless, colorless and
tasteless, so that its presence can not be readily detected until it is
too late.
		-- Chemical & Engineering News February 6, 1956