FORTUNE\'S PARTY TIPS		#14
Tired of finding that other people are helping themselves to your good
liquor at BYOB parties?  Take along a candle, which you insert and
light after you\'ve opened the bottle.  No one ever expects anything
drinkable to be in a bottle which has a candle stuck in its neck.