The astronomer Francesco Sizi, a contemporary of Galileo, argues that Jupiter can have no satellites:
	There are seven windows in the head, two nostrils, two ears, two
eyes, and a mouth; so in the heavens there are two favorable stars, two
unpropitious, two luminaries, and Mercury alone undecided and indifferent.
From which and many other similar phenomena of nature such as the seven
metals, etc., which it were tedious to enumerate, we gather that the number
of planets is necessarily seven. [...]
	Moreover, the satellites are invisible to the naked eye and
therefore can have no influence on the earth and therefore would be useless
and therefore do not exist.