Breathe deep the gathering gloom. Watch lights fade from every room. Bed-sitter people look back and lament; another day\'s useless energies spent.
Impassioned lovers wrestle as one.
Lonely man cries for love and has none.
New mother picks up and suckles her son.
Senior citizens wish they were young.

Cold-hearted orb that rules the night;
Removes the colors from our sight.
Red is grey and yellow white.
But we decide which is real, and which is an illusion.\
		-- The Moody Blues, \Days of Future Passed\