Fellow programmer, greetings\!  You are reading a letter which will bring you luck and good fortune.  Just mail (or UUCP) ten copies of this letter to ten of your friends.  Before you make the copies, send a chip or other bit of hardware, and 100 lines of \'C\' code to the first person on the list given at the bottom of this letter.  Then delete their name and add yours to the bottom of the list.
Don\'t break the chain\!  Make the copy within 48 hours.  Gerald R. of San
Diego failed to send out his ten copies and woke the next morning to find
his job description changed to \COBOL programmer.\  Fred A. of New York sent
out his ten copies and within a month had enough hardware and software to
build a Cray dedicated to playing Zork.  Martha H. of Chicago laughed at
this letter and broke the chain.  Shortly thereafter, a fire broke out in
her terminal and she now spends her days writing documentation for IBM PC\'s.

Don\'t break the chain\!  Send out your ten copies today\!
For example, if \thinmskip = 3mu, this makes \thickmskip = 6mu.  But if
you also want to use \skip12 for horizontal glue, whether in math mode or
not, the amount of skipping will be in points (e.g., 6pt).  The rule is
that glue in math mode varies with the size only when it is an \mskip;
when moving between an mskip and ordinary skip, the conversion factor
1mu=1pt is always used.  The meaning of \'\mskip\skip12\' and
\'\baselineskip=\the\thickmskip\' should be clear.
		-- Donald Knuth, TeX 82 -- Comparison with TeX80