Fortune\'s Rules for Memo Wars: #2
Given the incredible advances in sociocybernetics and telepsychology over
the last few years, we are now able to completely understand everything that
the author of an memo is trying to say.  Thanks to modern developments
in electrocommunications like notes, vnews, and electricity, we have an
incredible level of interunderstanding the likes of which civilization has
never known.  Thus, the possibility of your misinterpreting someone else\'s
memo is practically nil.  Knowing this, anyone who accuses you of having
done so is a liar, and should be treated accordingly.  If you *do* understand
the memo in question, but have absolutely nothing of substance to say, then
you have an excellent opportunity for a vicious ad hominem attack.  In fact,
the only *inappropriate* times for an ad hominem attack are as follows:

	1: When you agree completely with the author of an memo.
	2: When the author of the original memo is much bigger than you are.
	3: When replying to one of your own memos.