If Dr. Seuss Were a Technical Writer.....
Here\'s an easy game to play.
Here\'s an easy thing to say:

If a packet hits a pocket on a socket on a port,
And the bus is interrupted as a very last resort,
And the address of the memory makes your floppy disk abort,
Then the socket packet pocket has an error to report\!

If your cursor finds a menu item followed by a dash,
And the double-clicking icon puts your window in the trash,
And your data is corrupted \'cause the index doesn\'t hash,
then your situation\'s hopeless, and your system\'s gonna crash\!

You can\'t say this?  What a shame, sir\!
We\'ll find you another game, sir.

If the label on the cable on the table at your house,
Says the network is connected to the button on your mouse,
But your packets want to tunnel on another protocol,
That\'s repeatedly rejected by the printer down the hall,
And your screen is all distorted by the side effects of gauss,
So your icons in the window are as wavy as a souse,
Then you may as well reboot and go out with a bang,
\'Cause as sure as I\'m a poet, the sucker\'s gonna hang\!

When the copy of your floppy\'s getting sloppy on the disk,
And the microcode instructions cause unnecessary risc,
Then you have to flash your memory and you\'ll want to ram your rom.
Quickly turn off the computer and be sure to tell your mom\!

		-- DementDJ@ccip.perkin-elmer.com (DementDJ) [rec.humor.funny]