All the lines have been written		There\'s been Sandburg, It\'s sad but it\'s true			Keats, Poe and McKuen With all the words gone,		They all had their day What\'s a young poet to do?		And knew what they\'re doin\'
But of all the words written		The bird is a strange one,
And all the lines read,			So small and so tender
There\'s one I like most,		Its breed still unknown,
And by a bird it was said\!		Not to mention its gender.

It reminds me of days of		So what is this line
Both gloom and of light.		Whose author\'s unknown
It still lifts my spirits		And still makes me giggle
And starts the day right.		Even now that I\'m grown?

I\'ve read all the greats
Both starving and fat,
But none was as great as
\I tot I taw a puddy tat.\
		-- Etta Stallings, \An Ode To Childhood\