Patriotism is the last refuge of a scoundrel. 		-- S. Johnson, \The Life of Samuel Johnson\ by J. Boswell
In Dr. Johnson\'s famous dictionary patriotism is defined as the last
resort of the scoundrel.  With all due respect to an enlightened but
inferior lexicographer I beg to submit that it is the first.
		-- Ambrose Bierce

When Dr. Johnson defined patriotism as the last refuge of a scoundrel,
he ignored the enormous possibilities of the word reform.
		-- Sen. Roscoe Conkling

Public office is the last refuge of a scoundrel.
		-- Boies Penrose