Fortune\'s nomination for All-Time Champion and Protector of Youthful Morals goes to Representative Clare E. Hoffman of Michigan.  During an impassioned House debate over a proposed bill to \expand oyster and clam research,\ a sharp-eared informant transcribed the following exchange between our hero and Rep. John D. Dingell, also of Michigan.
DINGELL: There are places in the world at the present time where we are
	 having to artificially propagate oysters and clams.
HOFFMAN: You mean the oysters I buy are not nature\'s oysters?
DINGELL: They may or may not be natural.  The simple fact of the matter
	 is that female oysters through their living habits cast out
	 large amounts of seed and the male oysters cast out large
	 amounts of fertilization ...
HOFFMAN: Wait a minute\!  I do not want to go into that.  There are many
	 teenagers who read The Congressional Record.