Dear Emily, what about test messages? 		-- Concerned
Dear Concerned:
	It is important, when testing, to test the entire net.  Never test
merely a subnet distribution when the whole net can be done.  Also put \please
ignore\ on your test messages, since we all know that everybody always skips
a message with a line like that.  Don\'t use a subject like \My sex is female
but I demand to be addressed as male.\ because such articles are read in depth
by all USEnauts.
		-- Emily Postnews Answers Your Questions on Netiquette