But the greatest Electrical Pioneer of them all was Thomas Edison, who was a brilliant inventor despite the fact that he had little formal education and lived in New Jersey.  Edison\'s first major invention in 1877, was the phonograph, which could soon be found in thousands of American homes, where it basically sat until 1923, when the record was invented.  But Edison\'s greatest achievement came in 1879, when he invented the electric company. Edison\'s design was a brilliant adaptation of the simple electrical circuit: the electric company sends electricity through a wire to a customer, then immediately gets the electricity back through another wire, then (this is the brilliant part) sends it right back to the customer again.
This means that an electric company can sell a customer the same batch of
electricity thousands of times a day and never get caught, since very few
customers take the time to examine their electricity closely. In fact the
last year any new electricity was generated in the United States was 1937;
the electric companies have been merely re-selling it ever since, which is
why they have so much free time to apply for rate increases.
		-- Dave Barry, \What is Electricity?\